autogen id's

+++ mesteriApplication/src/main/java/com/intern/alexx/services/impl/MesterServiceImp.java
@@ -2,13 +2,15 @@ package com.intern.alexx.services.impl;
 
 import java.sql.SQLException;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.intern.alexx.model.Mester;
 import com.intern.alexx.model.MesterSearchCriteria;
-import com.intern.alexx.model.MyPage; 
+import com.intern.alexx.model.MyPage;
 import com.intern.alexx.repository.ContactRepository;
 import com.intern.alexx.repository.MesterRepository;
 import com.intern.alexx.repository.SpecialityRepository;
@@ -17,6 +19,8 @@ import com.intern.alexx.services.MesterService;
 @Component
 
 public class MesterServiceImp implements MesterService {
+	
+	private static Logger LOGGER = LoggerFactory.getLogger(MesterService.class);
 
 	@Autowired
 	private MesterRepository mesterRepository;
@@ -28,22 +32,30 @@ public class MesterServiceImp implements MesterService {
 	@Transactional
 	public void insertMester(Mester mester) {
 	 
+		if (mester.getId() != null){
+			throw new IllegalArgumentException("mester is ID be null");
+		}
 		mesterRepository.insert(mester);
-		System.out.println(mester.toString());
+		LOGGER.info(mester.toString());
 		contactRepository.insert(mester.getContact());
-		System.out.println(mester.getContact().toString());
+		LOGGER.info(mester.getContact().toString());
 		if (specialityRepository.getByName(mester.getSpeciality().getSpecialityName()) != null) {
 			mester.setSpeciality(specialityRepository.getByName(mester.getSpeciality().getSpecialityName()));
 		} else {
 			specialityRepository.insert(mester.getSpeciality());
-			System.out.println(mester.getSpeciality().toString());
+			LOGGER.info(mester.getSpeciality().toString());
 		}
 		mesterRepository.insertIntoMesterHasSpeciality(mester.getId(), mester.getSpeciality().getId());
-		System.out.println(mester.toString());
+		LOGGER.info(mester.toString());
 	}
 
 	@Transactional
 	public void updateMester(Mester mester) {
+		
+		if (mester.getId() == null){
+			throw new IllegalArgumentException("mester is ID must not be null");
+		}
+		
 		mesterRepository.deleteFromMesterHasSpeciality(mester.getId());
 		mesterRepository.update(mester);
 		contactRepository.update(mester.getContact());
@@ -58,6 +70,9 @@ public class MesterServiceImp implements MesterService {
 
 	@Transactional
 	public void deleteMester(String mesterId) {
+		if (mesterId != null){
+			throw new IllegalArgumentException("mester ID is null");
+		}
 		mesterRepository.deleteFromMesterHasSpeciality(mesterId);
 		contactRepository.delete(mesterId);
 		mesterRepository.delete(mesterId);
